Updated JavaFX UI/UX
