+++ src/main/java/duke/command/DeadlineCommand.java
@@ -37,11 +37,11 @@ public class DeadlineCommand extends Command {
                 Deadline deadline = new Deadline(deadlineDetails[0], date);
                 return taskList.addTask(deadline, storage);
             } else {
-                throw new DukeException("☹ OOPS!!! Please follow the syntax for an "
+                throw new DukeException("\uD83D\uDE14 OOPS!!! Please follow the syntax for an "
                         + "'deadline' command: event [description] /by [date].");
             }
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("☹ OOPS!!! Please follow the syntax for an "
+            throw new DukeException("\uD83D\uDE14 OOPS!!! Please follow the syntax for an "
                     + "'deadline' command: event [description] /by [date].");
         }
     }
