+++ src/main/java/duke/command/EventCommand.java
@@ -37,11 +37,11 @@ public class EventCommand extends Command {
                 Event event = new Event(eventDetails[0], date);
                 return taskList.addTask(event, storage);
             } else {
-                throw new DukeException("☹ OOPS!!! Please follow the syntax for an "
+                throw new DukeException("\uD83D\uDE14 OOPS!!! Please follow the syntax for an "
                         + "'event' command: event [description] /at [date].");
             }
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("☹ OOPS!!! Please follow the syntax for an "
+            throw new DukeException("\uD83D\uDE14 OOPS!!! Please follow the syntax for an "
                     + "'event' command: event [description] /at [date].");
         }
     }
