+++ src/main/java/duke/command/ExitCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.TaskList;
 import duke.utils.Storage;
-import duke.utils.Ui;
+import duke.gui.Ui;
 
 
 /**
