+++ src/main/java/duke/command/FindCommand.java
@@ -39,7 +39,7 @@ public class FindCommand extends Command {
             }
 
             if (taskIndexes.size() == 0) {
-                message = new StringBuilder("Sorry! I am unable to find any task with this keyword: "
+                message = new StringBuilder("\uD83D\uDE14 Sorry! I am unable to find any task with this keyword: "
                         + keyWord + "\n");
             } else {
                 for (Integer i: taskIndexes) {
@@ -50,7 +50,7 @@ public class FindCommand extends Command {
             return message.toString();
 
         } else {
-            throw new DukeException("☹ OOPS!!! Please only input ONE keyword.");
+            throw new DukeException("\uD83D\uDE14 OOPS!!! Please only input ONE keyword.");
         }
     }
 }
