+++ src/main/java/duke/command/UnmarkTaskCommand.java
@@ -31,14 +31,14 @@ public class UnmarkTaskCommand extends Command {
         try {
             taskList.get(index).markAsUndone();
             String message = "OK, I've marked this task as not done yet: \n  "
-                    + taskList.get(index) + "\n";
+                    + taskList.get(index);
             //Saving data
             storage.saveData(taskList.getList());
             return message;
         } catch (NullPointerException e) {
-            throw new DukeException("☹ OOPS!!! There is no task created for this index!");
+            throw new DukeException("\uD83D\uDE14 OOPS!!! There is no task created for this index!");
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("☹ OOPS!!! Please enter a valid index number!");
+            throw new DukeException("\uD83D\uDE14 OOPS!!! Please enter a valid index number!");
         }
     }
 }
