+++ src/main/resources/view/DialogBox.fxml
@@ -2,15 +2,27 @@
 
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.shape.Circle?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.HBox?>
 
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+<?import javafx.scene.text.TextFlow?>
 <fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" text="Label" wrapText="true" />
-        <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
+        <TextFlow fx:id="dialog">
+            <HBox.margin>
+                <Insets bottom="15.0" left="15.0" right="15.0" top="0.0"/>
+            </HBox.margin>
+            <padding>
+                <Insets bottom="15.0" left="15.0" right="15.0" top="15.0"/>
+            </padding>
+            <Text fx:id="text" style="-fx-font-family: Consolas; -fx-font-size: 12px; -fx-fill: white" />
+        </TextFlow>
+        <Circle fx:id="displayPicture" fill="WHITE" radius="25.0" stroke="#8899A6" strokeType="INSIDE" strokeWidth="1.0"/>
     </children>
     <padding>
-        <Insets bottom="15.0" left="5.0" right="5.0" top="15.0" />
+        <Insets bottom="15.0" left="10.0" right="10.0" top="15.0" />
     </padding>
 </fx:root>
